Switch to using a fixed mirror
There seems to be an issue with flathub builds not finding the source
when the central download link is used. This should still get us the
same source and is only slightly less reliable due to requiring that
this single mirror is up when the build is run.

This change can later be reverted when the issue with downloading from
that official URL in flathub build infra is fixed.