Update the launchable element in appdata
In the responses to issue #410 of flathub/flathub [1], it is suggested
that the value of the launchable element is not correct. Indeed in
appstream documentation [2] it is described as holding a desktop-file
id which in turn in simple cases is documented to basically be the
exact file name of the desktop file [3].

This commit just changes the launchable id to match the file name that
is used in the flatpak, org.gnu.emacs.desktop, which happens to differ
from the simple emacs.desktop name that is still used by
upstream. Ideally this should be cleaned up so that this patching is
not needed.

[1] https://github.com/flathub/flathub/issues/410
[2] https://www.freedesktop.org/software/appstream/docs/chap-Metadata.html#tag-launchable
[3] https://standards.freedesktop.org/desktop-entry-spec/desktop-entry-spec-latest.html#desktop-file-id